chore: 클래스 분리 및 코드정리
