+++ presentation/src/main/java/com/yapp/presentation/ui/admin/management/components/tablayout/YDSTabLayoutItemState.kt
@@ -1,33 +1,17 @@
 package com.yapp.presentation.ui.admin.management.components.tablayout
 
 
-data class YDSTabLayoutState(
-    val tabItems: List<YDSTabLayoutItemState>,
+data class YDSTabLayoutItemState(
+    val isSelected: Boolean,
+    val label: String,
 ) {
 
-    init {
-        require(tabItems.count { it.isSelected } == 1)
+    fun select(): YDSTabLayoutItemState {
+        return this.copy(isSelected = true)
     }
 
-    val selectedIndex: Int
-        get() = tabItems.indexOfFirst { it.isSelected }
-
-    fun select(targetIndex: Int): YDSTabLayoutState {
-        return this.copy(
-            tabItems = tabItems.mapIndexed { index, itemState ->
-                if (itemState.isSelected) {
-                    return@mapIndexed itemState.unSelect()
-                }
-
-                itemState
-            }.mapIndexed { index, itemState ->
-                if (index == targetIndex) {
-                    return@mapIndexed itemState.select()
-                }
-
-                itemState
-            }
-        )
+    fun unSelect(): YDSTabLayoutItemState {
+        return this.copy(isSelected = false)
     }
 
 }
